Make boot disks bigger instead of futzing with ramdisk and local ssd (#13079)
We keep running into disk space issues on the runners, especially with
larger framework-level models for benchmarking. I originally used a ram
disk because it was simple and fast, but with the current disk
requirements it's no longer simple and we're getting into the territory
of using too much RAM even on the runners that have tons of it. Local
ssd isn't a great option because with the n1-standard-96 machines we use
for builds, we'd have to attach 16 of them if we attach any. They also
require more configuration. Just making the boot disk big enough for
everything is simpler than what we have now, is easy to scale as far as
we should reasonably need to scale it, doesn't cost much, and doesn't
appear to have noticeable impact on build times.

Tested: deployed to test environment and ran CI for this pr on the test
runners, then manually compared CI job times with a recent postsubmit
run.


runner-env: testing