[xla] Drop arg & res attributes on XLA import (#13154)
During detupling the arg and res attributes end up failing to match. I
was initially going to flatten the arg/res attributes along with the
tuples but along the legacy XLA import path all the arg & result
attributes currently are not used, so just drop (some are related to
unneeded ABI, others to sharding which should be handled by partitioner
before here).