Add compiler C API `ireeCompilerGetRevision`. (#13184)
We also make use of the existing `ireeCompilerGetAPIVersion` (which
previously returned 0) to pack an API major/minor version. This is used
by the dynamic loader in order to conditionally load versioned symbols.

The API versioning is used to introduce the new entry point in a
compatible way such that it should not break existing deployments.