Add lowering for BF16 arith/math operations to F32 with truncf/extf wrappers (#12841)
We can support BF16 execution by casting to F32 types and performing the
F32 math in that space. This allows bf16 support once we have a
simplified conversion betweeen bf16 and f32. For now this path is
included after the BF16 to F32 promotion pass so it will not trigger. A
follow up will fully enable BF16 support and disable the F32 conversion.