Making iree_hal_channel_provider_t a ref object and exposing on devices. (#13317)
This allows for more complex provider implementations that may retain
state (like loaded dynamic libraries, initialization counters, etc) and
for providers to be shared across multiple devices. Via the new
iree_hal_device_replace_channel_provider API it's possible to now
configure channel providers after devices have been created instead of
needing to make the decisions in driver factory functions.

NCCL is now always loaded if available to decouple the need for a
channel provider to have been set at driver creation time. This also
allows for some devices to participate in collective operations and
others to not in a way that can be useful in hosting isolates that may
share the same driver.

OpenMPI isn't available on Windows so I got MSMPI working by adding
support for standard MPI implementations that use `int` instead of
pointers for various MPI types. We could extend this to be a cmake
setting as the heuristics are basically Windows/not-Windows but there
may be other things hiding that prevent us from wanting to do that.

Fixes #13303.