Remove the old low-level compiler C-API. (#13395)
In place of the old C-API, we add two new unstable interop functions
which allow the embedding API to be used to interface with the MLIR
C-API:

* ireeCompilerSessionGetContext
* ireeCompilerInvocationImportModule

Updates the compiler API test to demonstrate the mixed use of the
embedding API and the MLIR C-API for directly compiling an in-memory
module.

I wasn't going to prioritize doing this soon, but it turns out that it
is blocking further work on plugins, as it would force us to try to shim
IREE's plugin mechanism into generic use of the MLIR C-API.

This also removes the corresponding Python APIs which depended on these
features. The tests for these have been disabled and have been high cost
to keep working. We can bring back a better defined Python API if there
is ever demand.

Implements and fixes #13215.