Reduce need to specify input pipeline. (#13443)
Utilize dynamic pipeline to construct appropriate pipeline based on the
contents of the Module by default. This should be made even more
flexible in future (e.g., mixing the different input conversions), but
removes a footgun and improves user experience.

This does not cover every config (and this change does not removed
explicitly specified input conversion flags in tests), but should cover
reasonable set. The cost is an extra traversal and in future this could
be made even more dynamic (e.g., handling combinations of these
dialects). 

Related to #10297.