HLO input support for iree.abi.output and iree.abi.encoding attrs. (#13678)
This allows the original HLO types to be passed down to the later ABI
materialization passes even if the types are changed during conversion.
JAX currently emits a tf.aliasing_output arg attr to indicate when
outputs alias input memory and that's now passed down into the lower
layers of IREE to ensure we (try to) do the same.

This approach is an alternative to #13090 that avoids issues around
various sharp edges with signed types, func op legality, and the
potential for the unsigned types to survive to layers of the stack where
they are not expected.

Note that unfortunately this does not avoid the extra transient
allocations for aliased outputs yet (in most cases) as there are
improvements required in the `iree-stream-emplace-allocations` pass to
handle some of the ops. For now this at least drops the number of
external allocations to 0 but still needs the transients that get copied
into them.