Improve reduction code generation (#13688)
This patch emits fewer shuffle instructions than the previous approach.
We now only emit as many shuffle instructions as the dimensions of the
parallel loops. We accomplish this by first doing a reduction among all
the vector values within a thread and then only communicate this result
with the other threads.