Reimplement ukernel arch-specific code path fallbacks as weak symbols. (#13715)
This allows dropping `#if..#else` chains and seems like a step towards
multi-arch support in bitcode.

This relies on precise ordering of linked libraries, as the weak-symbol
fallback has to come last. In the current form of this PR, this is
achieved by separating the weak symbols into their own file `weak.c`,
built as its own separate `cc_library` named `zzz_weak_linklast` so that
it's alphabetically last so that it's kept at the end of `deps` by
`buildifier`. And as other deps from other packages (in particular, the
`arch/x86_64` etc. subpackages, which provide the non-fallback impls)
would still be ordered after that, the `ukernel` library is split into a
`ukernel_noweak` that has all those dependencies so that the final
`ukernel` library depends only on `ukernel_noweak` and
`zzz_weak_linklast`.

An earlier form of this PR relied on `-Wl,--whole-archive` and that was
worse still, build-wise.