Fix `bf16` cases involving `arith.bitcast` and `arith.constant` (#13733)
`arith.bitcast` is a cast where we do not wish to convert the `bf16`
type to a `f32` as bit conversion should
generically work, fixes #13722.

Converting `arith.constant` in the `bf16` to `ui16` buffer promotion
should raw convert the bits, not
perform an float-to-float cast. Fixes #13721.

This fixes also happen to fix #13720.