[CPU] Enforce 512-bit vector registers if target supports AVX-512 (#13754)
The default vector width used in LLVM for AVX-512 is 256 bits due to
frequency throttling issues when 512-bit vectors are used. However,
despite the throttling, 512-bit vectors should be more performant on
intensive computation scenarios (e.g., ML). This PR enforces 512-bit
vector when AVX-512 support is detected in the CPU target.

This should only impact the performance of generic ops as we were
already using 512-bit vector registers for matmuls and convolutions due
to the fixed tile sizes (e.g., 32 for the matmul vectorized dims or
`vectorSize * 2` for convolutions).