Drop conditionals and configured headers from the ukernels build (#13834)
This drop CMake `if()` and configured headers from the ukernels build.
Instead:

* The uarch-specific code paths are unconditionally added and then
`#if`'d out inside C code based on compiler version macros. This
simplifies the build ahead of multi-arch bitcode (#13804) and paves the
way for Bazel support on par with CMake.
* The `iree_uk_ssize_t` type is unconditionally `int64_t`, being more
clearly branded a size/index type more than a "pointer-size" type. This
wasn't completely clear until the recent bitcode push, but now that I'm
looking at LLVM IR where `ptr` is an opaque type and there isn't even a
concept of pointer size, while size/index have to use some explicitly
sized integer type, this change reflects the reality of what this type
is used for. It would be possible to make this `int32_t` instead of
`int64_t` for best performance on 32-bit architectures but for now
making it unconditionally `int64_t` is an easy first step and keeps what
we were currently doing in standalone builds.