Enable AVX2+FMA in e2e matmul + ukernels test; support comma-separated CPU features. (#13837)
The e2e matmul test covering llvmcpu-ukernels didn't exercise any
interesting architecture code path: since it didn't specify any CPU
features, it was falling back on the architecture-default, which, on
x86-64, would be SSE2, but we don't have a dedicated code path for that,
so it was falling back on scalar code.

This makes it use the AVX2+FMA path. Note, the trace-runner program used
by this trace-runner test, `iree-e2e-matmul-test.c`, will detect CPU
features and skip the test if the machine doesn't support these
features. Note, this is the true reason why this needs to be explicitly
labelled as "CPU features" and can't just be folded into copts --- we
need the CPU features called out explicitly in the yaml trace so the
trace runner can perform that check.

This required plumbing some support for multiple comma-separated CPU
features.

This is the last little stop on the route towards #13804 before we
finally bring in arm64 into the picture, making this officially
multi-arch...