Fix arm64 inline asm, was still referencing hardcoded register as in old out-of-line asm. (#13845)
In https://github.com/openxla/iree/pull/13383, when porting from
out-of-line to inline asm, I forgot to update the register operands to
refer to the named register inputs/outputs (e.g. `%[some_variable]`), I
left them as hardcoded registers (e.g. `x0`). This accidentally worked
as in the native build, the variables in question, being function
arguments, were still placed in the expected register as dictated by the
arm64 calling convention. But this broke down now as we are starting to
build this into bitcode!