Add --iree-codegen-linalg-max-constant-fold-elements= flag. (#13909)
Currently set to intmax so there's no changes to users models. Users can
opt in to disabling the extremely slow transpose folder as though it's
critical for reducing runtime memory consumption it is really bad for
iteration time on large models. Having a way to bound the folding is
useful in lieu of a real solution (fast transpose in C++, consteval,
whatever).

Workaround until #12033 (or another) does something better.