Moving builtins lower in the pipeline and adding option to force. (#13994)
By moving it after partitioning we can use the resource config local to
the partitioned execution region where the ops live to make decisions.
We also are able to preserve the semantics of the ops through
partitioning/placement such as whether splats should be rematerialized.

To work around the CUDA issue in #13984 a global flag was added to force
all splats/fills into builtins that must be passed when graphs are used
at runtime.

I was hoping this would make it easier to support `complex<f64>` as well
but that doesn't quite fit due to the lack of an `i128` in MLIR. One
day. May need to move the canonicalizeFillPattern doing the integer
conversion down into MaterializeBuiltins but that would mean removing
the verifier on AsyncSplatOp/AsyncFillOp that the type is actually
supported and spreads encoding across stages of the pipeline. Still may
be worth it for readability but we can re-evaluate when we have a need
for `complex<f64>`.